backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
google-auth
bcrypt
pydantic
python-multipart
sib-api-v3-sdk
